Nano Tach with IR LED/TRANS
Trying the simplest possible optical tach using an IR LED to transmit and an IR TRANsistor to either drive interrupts or to be polled. Experiments with both approaches. No other components except two resistors. See if it can be done without using OPAMPS or various level triggers. So So results but fun to try.